fix: cannot parse directory

+++ src/index.ts
@@ -38,23 +38,9 @@ export default async function transformEjsDir(source: string, target: string, op
     fileList.map(async (filePath) => {
       const fullPath = join(sourcePath, filePath);
       try {
-        // read
-        const file = await fs.readFile(fullPath);
-        let fileContent = file.toString();
-
+        const stat = await fs.stat(fullPath);
+        const targetDir = isAbsolute(target) ? target : join(cwd, target);
         let targetPath = filePath;
-        let shouldTransform = true;
-        if (ejsOnly) {
-          let ejsSuffix = typeof ejsOnly === 'string' ? ejsOnly : '.ejs';
-          if (!ejsSuffix.startsWith('.')) ejsSuffix = `.${ejsSuffix}`;
-          if (targetPath.endsWith(ejsSuffix)) {
-            targetPath = targetPath.slice(0, targetPath.length - ejsSuffix.length);
-          } else {
-            // do not transform
-            shouldTransform = false;
-          }
-        }
-
         // transform targetPath
         targetPath = ejs.render(
           targetPath,
@@ -78,28 +64,50 @@ export default async function transformEjsDir(source: string, target: string, op
           },
         );
 
-        if (shouldTransform) {
-          fileContent = ejs.render(fileContent, {
-            ...data,
-            camelCase,
-            pascalCase,
-            paramCase,
-            noCase,
-            dotCase,
-            capitalCase,
-            pathCase,
-            snakeCase,
-            headerCase,
-            sentenceCase,
-            constantCase,
-          });
-        }
+        if (stat.isDirectory()) {
+          const targetFullPath = join(targetDir, targetPath);
+          await fs.mkdir(targetFullPath, { recursive: true });
+        } else if (stat.isFile()) {
+          // read
+          const file = await fs.readFile(fullPath);
+          let fileContent = file.toString();
 
-        // write
-        const targetDir = isAbsolute(target) ? target : join(cwd, target);
-        const targetFullPath = join(targetDir, targetPath);
-        await fs.mkdir(dirname(targetFullPath), { recursive: true });
-        await fs.writeFile(targetFullPath, fileContent);
+          let shouldTransform = true;
+          if (ejsOnly) {
+            let ejsSuffix = typeof ejsOnly === 'string' ? ejsOnly : '.ejs';
+            if (!ejsSuffix.startsWith('.')) ejsSuffix = `.${ejsSuffix}`;
+            if (targetPath.endsWith(ejsSuffix)) {
+              targetPath = targetPath.slice(0, targetPath.length - ejsSuffix.length);
+            } else {
+              // do not transform
+              shouldTransform = false;
+            }
+          }
+
+          if (shouldTransform) {
+            fileContent = ejs.render(fileContent, {
+              ...data,
+              camelCase,
+              pascalCase,
+              paramCase,
+              noCase,
+              dotCase,
+              capitalCase,
+              pathCase,
+              snakeCase,
+              headerCase,
+              sentenceCase,
+              constantCase,
+            });
+          }
+
+          // write
+          const targetFullPath = join(targetDir, targetPath);
+          await fs.mkdir(dirname(targetFullPath), { recursive: true });
+          await fs.writeFile(targetFullPath, fileContent);
+        } else {
+          logger.warn(`Path [${fullPath}] ignored is not a file or directory.`)
+        }
       } catch (error: any) {
         logger.error('Error while transforming file:', fullPath);
         logger.error(error);
